requirements.txt
numpy 
mlflow
dagshub
xgboost
pandas
matplotlib
pymongo
python-dotenv
scikit-learn
boto
evidently

-e .